task #100

+++ 0x0C-more_malloc_free/100-realloc.c
@@ -8,9 +8,9 @@
  * Return: void
  */
 
-void _memcpy(void *dest, void *src, size_t n)
+void _memcpy(void *dest, void *src, unsigned int n)
 {
-	size_t i;
+	unsigned int i;
 	char *cdest = (char *)dest;
 	char *csrc = (char *)src;
